+++ app/Student.php
@@ -368,6 +368,21 @@ class Student extends Model implements HasMedia
                 }
             }
         }
+        if ($challenge->requirements) {
+            foreach ($challenge->requirements as $item) {
+                $studentItem = $this->items->where('id', $item['id'])->first();
+                if($studentItem)
+                 dump($studentItem->attributes   );
+                if ($studentItem)
+                    $count = $studentItem->pivot->count + $mult * -1;
+                else $count = 1;
+                if($count <= 0) {
+                    $this->items()->detach($item['id']);
+                } else {
+                    $this->items()->sync([$item['id'] => ['count' => $count]], false);
+                }
+            }
+        }
     }