+++ resources/lang/ca/auth.php
@@ -29,7 +29,7 @@ return [
     'confirm' => 'Confirma contrasenya',
     'username_info' => 'Ha de començar amb una lletra, 4 caràcters mínim i sense caràcters especials, espais...',
     'student' => 'Estudiant',
-    'teacher' => 'Docent',
-    'teacher_student' => 'Docent o estudiant?',
+    'teacher' => 'Profe',
+    'teacher_student' => 'Profe o estudiant?',
 
 ];