+++ resources/lang/es/auth.php
@@ -28,7 +28,7 @@ return [
     'confirm' => 'Confirma contraseña',
     'username_info' => 'Tiene que empezar con una letra, 4 caracteres mínimo y sin caracteres espciales, espacios...',
     'student' => 'Estudiante',
-    'teacher' => 'Docente',
-    'teacher_student' => '¿Docente o estudiante?',
+    'teacher' => 'Profe',
+    'teacher_student' => 'Profe o estudiante?',
 
 ];